Support validation of dialog contents; validate user-entered urls (#52)
use it to validate user-entered urls for images and links in dialogs

Note that our dialog implementation now has two separate notions of validation, one that occurs in `onChange` and one that occurs when the ok button is clicked. There may be a way to unify them at some point but I don't think that day is today.