Handle/ignore `className` property which confuses React (#53)
The symptom was a console error from React about an invalid `classNamename` property. I suspect the problem began with an import of invalid hand-coded HTML which looked something like `<ul><li className="foo">bar</li></ul>` which then introduced the `className` property into the system after which it was copied around to several other places. The fix here is to strip it out or convert it to `class` on import/export depending on circumstances.

Also adds some additional unit tests.