exports for a external variables plugin
Many of them are required because parts of the dialog system are woven into into
the built in toolbar. So these parts had to be reimplemented outside of slate-editor.
Probably the dialog system can be refactored so it is much easier to be used by systems
that provide their own toolbar.